Rename files from camelCase to kebab case
